+++ Cargo.toml
@@ -8,10 +8,11 @@ members = [
     "influxdb3_load_generator",
     "influxdb3_process",
     "influxdb3_server",
+    "influxdb3_telemetry",
+    "influxdb3_test_helpers",
     "influxdb3_wal",
     "influxdb3_write",
     "iox_query_influxql_rewrite",
-    "influxdb3_telemetry",
 ]
 default-members = ["influxdb3"]
