+++ influxdb3_test_helpers/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod object_store;
