+++ influxdb3_write/Cargo.toml
@@ -21,6 +21,7 @@ schema.workspace = true
 # Local deps
 influxdb3_catalog = { path = "../influxdb3_catalog" }
 influxdb3_id = { path = "../influxdb3_id" }
+influxdb3_test_helpers = { path = "../influxdb3_test_helpers" }
 influxdb3_wal = { path = "../influxdb3_wal" }
 
 # crates.io dependencies
